+++ x/bank/keeper/collections_test.go
@@ -26,6 +26,7 @@ import (
 
 func TestBankStateCompatibility(t *testing.T) {
 	key := storetypes.NewKVStoreKey(banktypes.StoreKey)
+	tkey := storetypes.NewTransientStoreKey(banktypes.TStoreKey)
 	testCtx := testutil.DefaultContextWithDB(t, key, storetypes.NewTransientStoreKey("transient_test"))
 	ctx := testCtx.Ctx.WithBlockHeader(cmtproto.Header{Time: cmttime.Now()})
 	encCfg := moduletestutil.MakeTestEncodingConfig()
@@ -44,6 +45,7 @@ func TestBankStateCompatibility(t *testing.T) {
 		map[string]bool{accAddrs[4].String(): true},
 		authtypes.NewModuleAddress("gov").String(),
 		log.NewNopLogger(),
+		tkey,
 	)
 
 	// test we can decode balances without problems